Feat: 음식 분류 모델 모듈화
- 환경변수 이름 수정
- README.md 수정

Related to: #5
