+++ Food_Classification/FoodClassifier.py
@@ -1,0 +1,44 @@
+import os, base64, requests
+
+# OpenAI API Key
+api_key = os.environ['OPENAI_API_KEY']
+
+def classify_food(image_path):
+  # Function to encode the image
+  def encode_image(image_path):
+    with open(image_path, "rb") as image_file:
+      return base64.b64encode(image_file.read()).decode('utf-8')
+
+  # Getting the base64 string
+  base64_image = encode_image(image_path)
+
+  headers = {
+    "Content-Type": "application/json",
+    "Authorization": f"Bearer {api_key}"
+  }
+
+  payload = {
+    "model": "gpt-4o",
+    "messages": [
+      {
+        "role": "user",
+        "content": [
+          {
+            "type": "text",
+            "text": "What’s this food? just say name in Korean"
+          },
+          {
+            "type": "image_url",
+            "image_url": {
+              "url": f"data:image/png;base64,{base64_image}"
+            }
+          }
+        ]
+      }
+    ],
+    "max_tokens": 300
+  }
+
+  response = requests.post("https://api.openai.com/v1/chat/completions", headers=headers, json=payload)
+
+  return response.json()['choices'][0]['message']['content']
