--- Food_Classification/Food_Classification_with_OpenAI_API.py
@@ -1,46 +1,0 @@
-import os, base64, sys, requests
-
-# OpenAI API Key
-api_key = os.environ['GPT_API_KEY']
-
-# Function to encode the image
-def encode_image(image_path):
-  with open(image_path, "rb") as image_file:
-    return base64.b64encode(image_file.read()).decode('utf-8')
-
-# Path to your image
-image_path = sys.argv[1]
-
-# Getting the base64 string
-base64_image = encode_image(image_path)
-
-headers = {
-  "Content-Type": "application/json",
-  "Authorization": f"Bearer {api_key}"
-}
-
-payload = {
-  "model": "gpt-4o",
-  "messages": [
-    {
-      "role": "user",
-      "content": [
-        {
-          "type": "text",
-          "text": "What’s this food? just say name"
-        },
-        {
-          "type": "image_url",
-          "image_url": {
-            "url": f"data:image/png;base64,{base64_image}"
-          }
-        }
-      ]
-    }
-  ],
-  "max_tokens": 300
-}
-
-response = requests.post("https://api.openai.com/v1/chat/completions", headers=headers, json=payload)
-
-print(response.json()['choices'][0]['message']['content'])
